New gulp
Oh boy. A lot of small changes, many performance fixes, and reworked some workflows.
It is now possible to use resolve to find the absolute path of a image. This is the new required way of using images in SASS files.

It is now possible to use import and other ES6 functions in JS files. Also usage of NPM modules is now allowed.